feat: update insert_remove to avoid cloning the returned key
The insert_remove method previously cloned the key in some situations
to allow returning an owned K when only &K was available. Using a Cow
pointer, it is possible to return an owned or borrowed K depending on
whether ownership is still needed by the cache. This avoids cloning
entirely within the implementation, though the trait bound "K: Clone"
is still required due to the trait bound on Cow pointers.